+++ PersistentStorage/Sources/Data/Repository/PersistentStorage/MainPersistentStorage.swift
@@ -1,71 +1,7 @@
-import Foundation
 import Combine
+import Foundation
 
-public protocol PersistentStorage: PersistentStorageCodable, PersistentStorageDeprecated {
-    /// Method for storing a value with assigned key into the user defaults storage.
-    ///
-    /// - Parameters:
-    ///     - value: The value you want to store in
-    ///     - key: Key that will be assigned to value
-    @discardableResult
-    func store<T>(
-        _ value: T?,
-        for key: String
-    ) -> AnyPublisher<Void, PersistentStorageError>
-
-    /// Method for removing a value with assigned key from the user defaults storage.
-    ///
-    /// - Parameters:
-    ///     - key: Key that will be assigned to value
-    @discardableResult
-    func remove(
-        valueKey: String
-    ) -> AnyPublisher<Void, PersistentStorageError>
-
-    /// Method for reading a value with given key and type.
-    ///
-    /// - Possible failures:
-    ///     - If type resolution with given parameter `valueType` fails, method will return `noValueFoundWithGivenType` failure.
-    ///     - If no value is associated with key given in parameter `valueKey` or its value is nil, the method will return `noValueFound` failure.
-    ///
-    /// - Parameters:
-    ///     - valueType: The type of value that you wish to read.
-    ///     - valueKey: Key that is assigned to a value that you wish to read.
-    func read<T>(
-        valueType: T.Type,
-        valueKey: String
-    ) throws -> T
-
-    /// Method for reading a value with given key and type. Result is returned as a publisher.
-    ///
-    /// - Possible failures:
-    ///     - If type resolution with given parameter `valueType` fails, method will return `noValueFoundWithGivenType` failure.
-    ///     - If no value is associated with key given in parameter `valueKey` or its value is nil, the method will return `noValueFound` failure.
-    ///
-    /// - Parameters:
-    ///     - valueType: The type of value that you wish to read.
-    ///     - valueKey: Key that is assigned to a value that you wish to read.
-    func readWithPublisher<T>(
-        valueType: T.Type,
-        valueKey: String
-    ) -> AnyPublisher<T, PersistentStorageError>
-
-    /// Method for observing a value with given user defaults `KeyPath`.
-    ///
-    /// - Possible failures:
-    ///     - If no value is associated with key given in parameter `valueKey` or its value is nil, the method will return `noValueFound` failure.
-    ///
-    /// - Parameters:
-    ///     - keyPath: Specified key represented by user defaults `KeyPath` value.
-    func observe<T>(
-        keyPath: KeyPath<UserDefaults, T?>
-    ) -> AnyPublisher<T, PersistentStorageError>
-}
-
-/// The `PersistentStorage` class wraps apple native `UserDefaults` class and  provides a custom
-/// programmatic interface for complex accessing user defaults database.
 public class MainPersistentStorage: PersistentStorage {
-
     let userDefaults = UserDefaults.standard
     private let logConfiguration: LogConfiguration
